ref: move winner classes to models
