+++ src/entities/winners.ts
@@ -1,16 +1,4 @@
-export interface IWinner {
-  name: string;
-  wins: number;
-}
-
-export class Winner implements IWinner {
-  name: string = '';
-  wins: number = 0;
-
-  constructor(winner: Partial<Winner>) {
-    Object.assign(this, winner);
-  }
-}
+import { Winner } from '../models';
 
 export class WinnersList {
   private winners: Winner[] = [];
