+++ src/models/index.ts
@@ -1,2 +1,3 @@
 export * from './message.model';
 export * from './user.models';
+export * from './winner.models';
